Fix bug in update_student_extensions where wrong user_id is used.
The `user_id` must be the "gradebook" user id which is different from
the `ssid`, lms user id, and the id number currently recorded in
the `Member` dataclass. This commit gathers this `gradebook_user_id`
when extracting a member list from the roster page and uses
this `gradebook_user_id` when calling `update_student_extensions`.

Some convenience CLI tools have been added to expose this functionality.